Move utilities tests to corresponding packages for test coverage
