+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>org.hygge</groupId>
     <artifactId>hyggemvc</artifactId>
-    <version>0.8.0-beta</version>
+    <version>0.8.1-beta</version>
 
     <properties>
         <!-- github server corresponds to entry in ~/.m2/settings.xml -->
